Refactoring interfaces to remain consistent and one source of truth
