+++ client/src/interface/KanbanContext.ts
@@ -1,7 +1,7 @@
 import { DropResult } from 'react-beautiful-dnd';
 import { IBoard, NewBoardApiData, ICard, IColumn } from './';
 
-export interface IKanbanContext {
+export default interface IKanbanContext {
   activeBoard: IBoard;
   focusedCard: ICard | null;
   userBoards: IBoard[];
