+++ client/src/interface/index.ts
@@ -1,10 +1,11 @@
 export type { IUser, ISearchUsersApiData } from './User';
 export type { IBoard, NewBoardApiData, BoardApiData } from './Board';
-export type { ICard } from './Card';
-export type { IColumn } from './Column';
+export type { default as ICard } from './Card';
+export type { default as IColumn } from './Column';
 export type { IDialogItem } from './DialogItem';
-export type { IFetchOptions } from './FetchOptions';
+export type { default as IFetchOptions } from './FetchOptions';
 export type { IImageUploadData, IUploadOptions } from './File';
 export type { IAuthApiDataSuccess, IAuthApiData } from './AuthApiData';
-export type { IKanbanContext } from './KanbanContext';
-export type { IDialogItemContext } from './DialogItemContext';
+export type { default as IKanbanContext } from './KanbanContext';
+export type { default as IDialogItemContext } from './DialogItemContext';
+export type { default as INotificationItem } from './Notification';
